Implemented container and presentational components
