+++ client/src/components/App.js
@@ -1,0 +1,85 @@
+import { bindActionCreators } from 'redux';
+import { connect } from 'react-redux';
+import * as actionCreators from '../actions/action_creators.jsx';
+import Main from './Main.jsx';
+
+
+function mapStateToProps(state){
+  return {
+    header: state.header,
+    tournament: state.tournament,
+    mode: state.mode,
+  }
+}
+
+function mapDispatchToProps(dispatch){
+  return bindActionCreators(actionCreators, dispatch);
+}
+
+const App = connect(mapStateToProps, mapDispatchToProps)(Main);
+console.log("sup")
+export default App;
+
+/////////////////////////////////////////////////////////////////
+// //Pres...
+// import React, { PropTypes } from 'react'
+// import Todo from './Todo'
+//
+// const TodoList = ({ todos, onTodoClick }) => (
+//   <ul>
+//     {todos.map(todo =>
+//       <Todo
+//         key={todo.id}
+//         {...todo}
+//         onClick={() => onTodoClick(todo.id)}
+//       />
+//     )}
+//   </ul>
+// )
+//
+// TodoList.propTypes = {
+//   todos: PropTypes.arrayOf(PropTypes.shape({
+//     id: PropTypes.number.isRequired,
+//     completed: PropTypes.bool.isRequired,
+//     text: PropTypes.string.isRequired
+//   }).isRequired).isRequired,
+//   onTodoClick: PropTypes.func.isRequired
+// }
+//
+// export default TodoList
+// //Contain...
+// import { connect } from 'react-redux'
+// import { toggleTodo } from '../actions'
+// import TodoList from '../components/TodoList'
+//
+// const getVisibleTodos = (todos, filter) => {
+//   switch (filter) {
+//     case 'SHOW_ALL':
+//       return todos
+//     case 'SHOW_COMPLETED':
+//       return todos.filter(t => t.completed)
+//     case 'SHOW_ACTIVE':
+//       return todos.filter(t => !t.completed)
+//   }
+// }
+//
+// const mapStateToProps = (state) => {
+//   return {
+//     todos: getVisibleTodos(state.todos, state.visibilityFilter)
+//   }
+// }
+//
+// const mapDispatchToProps = (dispatch) => {
+//   return {
+//     onTodoClick: (id) => {
+//       dispatch(toggleTodo(id))
+//     }
+//   }
+// }
+//
+// const VisibleTodoList = connect(
+//   mapStateToProps,
+//   mapDispatchToProps
+// )(TodoList)
+//
+// export default VisibleTodoList
