+++ client/src/components/Main.jsx
@@ -6,7 +6,7 @@ import { Provider } from 'react-redux';
 import store from '../store.jsx';
 import Chat from './Chat.jsx';
 import { Grid, Row, Col } from 'react-bootstrap';
-import App from './App.jsx';
+
 
 // import io from 'socket.io-client';
 /* class MainComponent extends React.Component {
@@ -43,3 +43,23 @@ const Main = React.createClass({
 });
 
 export default Main;
+/////////////////////////////////////////////////////////////////
+// var cloneWithProps = require('react-addons-clone-with-props');
+//
+// var _makeBlue = function(element) {
+//   return cloneWithProps(element, {style: {color: 'blue'}});
+// };
+//
+// var Blue = React.createClass({
+//   render: function() {
+//     var blueChildren = React.Children.map(this.props.children, _makeBlue);
+//     return <div>{blueChildren}</div>;
+//   }
+// });
+//
+// ReactDOM.render(
+//   <Blue>
+//     <p>This text is blue.</p>
+//   </Blue>,
+//   document.getElementById('container')
+// );
