+++ webpack.config.js
@@ -14,7 +14,7 @@ module.exports = {
   output: { path: '/', filename: 'app-bundle.js' },
   module: {
     loaders: [{
-      test: /\.jsx$/,
+      test: /\.js|x$/,
       exclude: /node_modules/,
       loader: 'babel',
     },
